+++ ent/name/name.go
@@ -13,7 +13,7 @@ import (
 type DetectedName struct {
 	PageID         string    `gorm:"type:varchar(255);index"`
 	ItemID         int       `gorm:"not null;index"`
-	NameString     string    `sql:"type:CHARACTER VARYING(255) COLLATE \"C\" NOT NULL" gorm:"index"`
+	Name           string    `sql:"type:CHARACTER VARYING(255) COLLATE \"C\" NOT NULL" gorm:"index"`
 	AnnotNomen     string    `gorm:"type:varchar(255)"`
 	AnnotNomenType string    `gorm:"type:varchar(255)"`
 	OffsetStart    int       `gorm:"not null;default:0"`
@@ -25,26 +25,26 @@ type DetectedName struct {
 }
 
 type VerifiedName struct {
-	ID                int    `gorm:"primary_key"`
-	Name              string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\" NOT NULL"`
-	RecordID          string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\"" gorm:"index"`
-	MatchType         string `gorm:"type:varchar(100)"`
-	EditDistance      int    `gorm:"not null;default:0"`
-	StemEditDistance  int    `gorm:"not null;default:0"`
-	MatchedName       string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
-	MatchedCanonical  string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
-	CurrentName       string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
-	CurrentCanonical  string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
-	Classification    string `sql:"type:CHARACTER VARYING COLLATE \"C\""`
-	DataSourceID      int
-	DataSourceTitle   string `sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
-	DataSourcesNumber int
-	Curation          string
-	OddsLog10         float64 `gorm:"type:float;not null;default:0"`
-	Occurrences       int     `gorm:"not null;default:0"`
-	Retries           int     `gorm:"not null;default:0"`
-	Error             string
-	UpdatedAt         time.Time `sql:"type:timestamp without time zone"`
+	ID                int       `json:"id" gorm:"primary_key"`
+	Name              string    `json:"name" sql:"type:CHARACTER VARYING(255) COLLATE \"C\" NOT NULL"`
+	RecordID          string    `json:"recordID" sql:"type:CHARACTER VARYING(255) COLLATE \"C\"" gorm:"index"`
+	MatchType         string    `json:"matchType" gorm:"type:varchar(100)"`
+	EditDistance      int       `json:"editDistance" gorm:"not null;default:0"`
+	StemEditDistance  int       `json:"stemEditDistance" gorm:"not null;default:0"`
+	MatchedName       string    `json:"matchedName" sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
+	MatchedCanonical  string    `json:"matchedCanonical" sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
+	CurrentName       string    `json:"currentName" sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
+	CurrentCanonical  string    `json:"currentCanonical" sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
+	Classification    string    `json:"classification" sql:"type:CHARACTER VARYING COLLATE \"C\""`
+	DataSourceID      int       `json:"dataSourceID" gorm:"index"`
+	DataSourceTitle   string    `json:"dataSourceTitle" sql:"type:CHARACTER VARYING(255) COLLATE \"C\""`
+	DataSourcesNumber int       `json:"dataSourcesNumber"`
+	Curation          string    `json:"curation"`
+	OddsLog10         float64   `json:"oddsLog10" gorm:"type:float;not null;default:0"`
+	Occurrences       int       `json:"occurrences" gorm:"not null;default:0"`
+	Retries           int       `json:"-" gorm:"not null;default:0"`
+	Error             string    `json:"error"`
+	UpdatedAt         time.Time `json:"updatedAt" sql:"type:timestamp without time zone"`
 }
 
 type UniqueName struct {
@@ -66,7 +66,7 @@ func New(itemID int, p *page.Page, n gnfout.Name) DetectedName {
 	dn := DetectedName{
 		PageID:         p.ID,
 		ItemID:         itemID,
-		NameString:     n.Name,
+		Name:           n.Name,
 		OffsetStart:    start,
 		OffsetEnd:      end,
 		AnnotNomen:     n.AnnotNomen,