+++ ent/page/page.go
@@ -1,8 +1,8 @@
 package page
 
 type Page struct {
-	ID         string `gorm:"type:varchar(255);primary_key;auto_increment:false"`
-	ItemID     int    `gorm:"primary_key;auto_increment:false"`
-	Offset     int
-	OffsetNext int `gorm:"-"`
+	ID         string `json:"id" gorm:"type:varchar(255);primary_key;auto_increment:false"`
+	ItemID     int    `json:"item_id" gorm:"primary_key;auto_increment:false"`
+	Offset     int    `json:"-"`
+	OffsetNext int    `json:"-" gorm:"-"`
 }