+++ interface.go
@@ -1,6 +1,7 @@
 package bhlindex
 
 import (
+	"github.com/gnames/bhlindex/config"
 	"github.com/gnames/bhlindex/ent/finder"
 	"github.com/gnames/bhlindex/ent/loader"
 	"github.com/gnames/bhlindex/ent/verif"
@@ -11,4 +12,5 @@ type BHLindex interface {
 	FindNames(loader.Loader, finder.Finder) error
 	VerifyNames(verif.VerifierBHL) error
 	GetVersion() gnvers.Version
+	GetConfig() config.Config
 }